Test refactoring and shuffle
